Rename fixtures to mocks
